basic quiz with context
