+++ client/src/pages/Quiz/Question.jsx
@@ -1,25 +1,22 @@
 // ---- hooks, dependencies, styling import ----
-import React from 'react'
-
-// ---- components ----
+import { useContext } from "react";
+import "./animation.scss";
 
 // ---- context import ----
-
-// ---- data ----
+import QuizContext from "../../contexts/QuizContext/QuizContext";
 
 // ---- COMPONENT ----
 
-const Question = ({
-  questions,
-  currentQuestion,
-  handleAnswer,
-  answers,
-  chooseAnswer,
-  notify
-}) => {
-
-
-
+const Question = () => {
+  //? ---- hooks ----
+  const {
+    handleAnswer,
+    currentQuestion,
+    answers,
+    chooseAnswer,
+    notify,
+    questions,
+  } = useContext(QuizContext);
 
   //? ---- rendering ----
   return (
@@ -37,7 +34,9 @@ const Question = ({
         <div className=" h-56 flex flex-col items-center justify-around">
           {answers.map((answer, i) => (
             <button
-              className={`border-2 rounded-full text-stone-700 text-amber-50 font-bold py-2 px-4 w-60 focus:ring focus:ring-cyan-600 hover:bg-stone-100 focus:bg-stone-100 ${notify ? "wobble-hor-bottom" : null}` }
+              className={`border-2 rounded-full text-stone-700 text-amber-50 font-bold py-2 px-4 w-60 focus:ring focus:ring-cyan-600 hover:bg-stone-100 focus:bg-stone-100 ${
+                notify ? "wobble-hor-bottom" : null
+              }`}
               onClick={chooseAnswer}
               key={i}
             >
