+++ client/src/pages/Quiz/Quiz.jsx
@@ -1,6 +1,6 @@
 // ---- hooks, dependencies, styling import ----
-import './animation.scss'
-import { useState, useEffect } from "react";
+import { useContext } from "react";
+import "./animation.scss";
 
 // ---- components ----
 import Question from "./Question";
@@ -8,60 +8,13 @@ import Answer from "./Answer";
 import Results from "./Results";
 
 // ---- context import ----
-
-// ---- data ----
-import data from "./desert.json";
+import QuizContext from "../../contexts/QuizContext/QuizContext";
 
 // ---- COMPONENT ----
 
 const Quiz = ({ biomeName }) => {
   //? ---- hooks ----
-  const [currentQuestion, setCurrentQuestion] = useState(0);
-  const [showResults, setShowResults] = useState(false);
-  const [isSubmitted, setIsSubmitted] = useState(false);
-  const [selectedAnswer, setSelectedAnswer] = useState("");
-  const [checkAnswer, setCheckAnswer] = useState(false);
-  const [score, setScore] = useState(0);
-  const [notify, setNotify] = useState(false);
-
-  //? ---- data----
-  const { questions } = data;
-  const { correct_answer, answers, message } = questions[currentQuestion];
- 
-
-  //? ---- event handlers ----
-
-  const animate = () => {
-    setNotify(true);
-    setTimeout(()=>setNotify(false), 1000)
-  }
-
-  const handleAnswer = () => {
-    selectedAnswer === ""
-      ? animate()
-      : setIsSubmitted(true);
-    if (selectedAnswer === correct_answer) {
-      setCheckAnswer(true);
-      setScore(score + 1);
-    } else {
-      setCheckAnswer(false);
-    }
-    setSelectedAnswer("");
-  };
-
-  const handleNextQuestion = () => {
-    setIsSubmitted(false);
-    const nextQuestion = currentQuestion + 1;
-    if (nextQuestion < questions.length) {
-      setCurrentQuestion(nextQuestion);
-    } else {
-      setShowResults(true);
-    }
-  };
-
-  const chooseAnswer = (e) => {
-    setSelectedAnswer(e.target.textContent);
-  };
+  const { showResults, isSubmitted } = useContext(QuizContext);
 
   //? ---- rendering ----
   return (
@@ -70,26 +23,11 @@ const Quiz = ({ biomeName }) => {
         Can you help the Scientist write a chapter about the {biomeName}?
       </p>
       {showResults ? (
-        <Results biomeName={biomeName} score={score} questions={questions} />
+        <Results biomeName={biomeName} />
       ) : isSubmitted ? (
-        <Answer
-          checkAnswer={checkAnswer}
-          correct_answer={correct_answer}
-          message={message}
-          handleNextQuestion={handleNextQuestion}
-          currentQuestion={currentQuestion}
-          questions={questions}
-        />
+        <Answer />
       ) : (
-        <Question
-          questions={questions}
-          currentQuestion={currentQuestion}
-          handleAnswer={handleAnswer}
-          handleNextQuestion={handleNextQuestion}
-          answers={answers}
-          chooseAnswer={chooseAnswer}
-          notify={notify}
-        />
+        <Question />
       )}
     </div>
   );
